Makes use_conda off by default.
Building a separate conda environment for each test is orders of
magnitude slower then using a default conda environment. I am going to
turn this option off by default and only use it for packages that need
py27.

I think the slow down is caused by lcdblib not being in conda. lcdblib
has biopython/pybedtools as requirements, so when install with pip it is
building these packages which take a lot of time. Once we push a release
we may want to set this back to true.